Update Delete button, exception current user null

+++ HRM.UI/ViewModels/AddFamilyInforViewModel.cs
@@ -180,6 +180,13 @@ namespace HRM.UI.ViewModels
             {
                 if (string.IsNullOrEmpty(HoVaTen))
                     return false;
+
+                if (_userStore.CurrentNhanSu?.MaNhanVien == null) 
+                {
+                    MessageBox.Show("Chưa nhập mã nhân viên.", "Thông báo", MessageBoxButton.OK, MessageBoxImage.Warning);
+                    return false;
+                }
+
                 return true;
             }, async (p) =>
 
@@ -204,18 +211,19 @@ namespace HRM.UI.ViewModels
                     LoadData();
                     if (QuanHeGiaDinh != null)
                     {
-                        MessageBox.Show("Thêm thành công");
+                        MessageBox.Show("Thêm thành công", "Thông báo", MessageBoxButton.OK, MessageBoxImage.Information);
                         LoadData();
                     }
                     else
                     {
-                        MessageBox.Show("Lỗi hệ thống");
+                        MessageBox.Show("Lỗi hệ thống", "Lỗi", MessageBoxButton.OK, MessageBoxImage.Error);
                     }
 
                 }
                 catch (Exception ex)
                 {
                     await _unitOfWork.RollbackAsync();
+                    MessageBox.Show($"Đã xảy ra lỗi: {ex.Message}", "Lỗi", MessageBoxButton.OK, MessageBoxImage.Error);
                 }
             });
 
@@ -226,20 +234,33 @@ namespace HRM.UI.ViewModels
                 return true;
             }, async (p) =>
             {
-                await _unitOfWork.BeginTransactionAsync();
-                try
+                if (SelectedItem != null)
                 {
-                    var QuanHeGiaDinh = await _quanHeGiaDinhRepository.AsQueryable().FirstOrDefaultAsync(x => x.Id == SelectedItem.Id);
-                    await _quanHeGiaDinhRepository.DeleteAsync(QuanHeGiaDinh);
-                    await _unitOfWork.CommitAsync();
-                    LoadData();
+                    MessageBoxResult result = MessageBox.Show("Bạn có chắc chắn muốn xóa?", "Xác nhận xóa", MessageBoxButton.OKCancel, MessageBoxImage.Warning);
+
+                    if (result == MessageBoxResult.OK) 
+                    {
+                        await _unitOfWork.BeginTransactionAsync();
+
+                        try
+                        {
+                            var QuanHeGiaDinh = await _quanHeGiaDinhRepository.AsQueryable().FirstOrDefaultAsync(x => x.Id == SelectedItem.Id);
+
+                            if (QuanHeGiaDinh != null)
+                            {
+                                await _quanHeGiaDinhRepository.DeleteAsync(QuanHeGiaDinh);
+                                await _unitOfWork.CommitAsync();
+                                LoadData();
+                            }
+                        }
+                        catch (Exception ex)
+                        {
+                            await _unitOfWork.RollbackAsync();
+                            MessageBox.Show($"Đã xảy ra lỗi: {ex.Message}", "Lỗi", MessageBoxButton.OK, MessageBoxImage.Error);
+                        }
+                    }
                 }
-                catch (Exception ex)
-                {
-                    await _unitOfWork.RollbackAsync();
-                }
-            }
-            );
+            });
         }
         public void LoadData()
         {
